fixed #10649 La libreria password_compact che replicava la funzione password_hash, gestisce i parametri in modo diverso. Ora corretto.

+++ html/lib/lib.password.php
@@ -32,7 +32,7 @@ class Password {
     function Password($password) {
         $this->policies=PasswordPolicies::check($password);
         $this->password=$password;
-        $this->algorithm_default=Get::sett('pass_algorithm', PASSWORD_MD5);
+        $this->algorithm_default=(int)Get::sett('pass_algorithm', PASSWORD_MD5);
         $this->algorithm_options = array(
             PASSWORD_BCRYPT=>array(
                 'cost'=>10
